qemu: Move shareable disk check for block copy
The writing to an image actually starts when the copy job is initiated,
so checking this at the time of the pivot operation is too late.

Move the check to qemuDomainBlockCopyCommon. Note that modern qemu would
have prevented two writers with qcow2 so the slim possibility of a job
started with libvirtd without this patch missing the check is not really
worth worrying about.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <a31b462bec66c254d7693b325e6a5e6aa5c06bf2.1548942191.git.pkrempa@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>