qemu: caps: Use CAP_DAC_OVERRIDE for probing to avoid permission issues
This is mainly about /dev/sev and its default permissions 0600. Of
course, rule of 'tinfoil' would be that we can't trust anything, but the
probing code in QEMU is considered safe from security's perspective + we
can't create an udev rule for this at the moment, because ioctls and
filesystem permisions are cross checked in kernel and therefore a user
with read permisions could issue a 'privileged' operation on SEV which
is currently only limited to root.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <6acde6faba93fd1d12581c984e0f846c78969eb0.1548948096.git.eskultet@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>