qemu: capabilities: Probe caps for 'ide-hd' instead of 'ide-drive'
Since commit a4cda054e7 we are using 'ide-hd' and 'ide-cd' instead of
'ide-drive'. We also should probe capabilities for 'ide-hd' instead of
'ide-drive'. It is safe to do as 'ide-drive' is the common denominator
of both 'ide-hd' and 'ide-cd' so all the properties were common.

For now the test data are modified by just changing the appropriate type
 when probing for caps.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <b2e2a359c9db4f86ff43c883a1a2fda34209e953.1549295166.git.pkrempa@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>