qemu: caps: Introduce capability for 'device_id' property of 'scsi-disk'
The property allows to control the guest-visible content of the vendor
specific designator of the 'Device Identification' page of a SCSI
device's VPD (vital product data).

QEMU was leaking the id string of -drive as the value if the 'serial' of
the disk was not specified. Switching to -blockdev would impose an ABI
change.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <7c2a07141651017191c086b590ccc52c3362274e.1549295166.git.pkrempa@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>