wip: backup: Wire up qemu checkpoint commands over QMP
Time to actually issue the QMP transactions that create and
delete persistent checkpoints.  For create, we only need one
transaction: inside, we visit all disks affected by the
checkpoint, and create a new enabled bitmap, as well as
disabling the bitmap of the parent checkpoint (if any).  For
deletion, we need multiple calls: if the checkpoint to be
deleted is active, we must enable the parent; then we must
merge the existing checkpoint into the parent, and finally
we can delete the checkpoint.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190206191818.14646-18-eblake@redhat.com>