qemuDomainDetachChrDevice: Use @tmpChr to build device string
So far we are passing @chr to qemuBuildChrDeviceStr. This is
suboptimal beacuse @chr is just parsed XML definition provided by
user which by definition may lack some information. On the other
hand, @tmpChr is the one that was found using @chr in domain
definition so it contains the same amount of information or more.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <ec473b184c56d26ceb3361eb2a7770bccefaff32.1549899626.git.mprivozn@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>