domain: Define explicit flags for saved image xml
Commit d2a929d4 (0.9.4) defined virDomainSaveImageGetXMLDesc()'s use
of @flags as a subset of virDomainXMLFlags, documenting that 2 of the
3 flags defined at the time would never be valid.  Later, commit
28f8dfdc (1.0.0) introduced a new flag, VIR_DOMAIN_XML_MIGRATABLE, but
did not adjust the save image documentation to declare it as invalid.
Later, commit a67e3872 (3.7.0) blindly copied and pasted the same text
into virDomainManagedSaveGetXMLDesc.

However, since the flag is not accepted as valid by any of the
drivers (remote is just passthrough; and qemu is the only supporting
driver with support for just VIR_DOMAIN_XML_SECURE), and it is
unlikely that the domain state saved off during saved image creation
needs to be migration-friendly (the saved image is loaded by the
current libvirt version, and not migrated to a potentially older
version), it is easier to just define an explicit set of supported
flags directly related to the snapshot API rather than trying to
borrow from live domain API, and risking confusion if even more
domain flags are added later (in fact, I have an upcoming patch that
plans to add a new flag to virDomainGetXMLDesc that makes no sense for
saved images).  On the other hand, it DOES make sense to reuse the
same flags for SaveImage and for ManagedSave (since ManagedSave is
really just sugar for creating a normal SaveImage in a location
controlled by libvirt instead of by the user).

There is no API or ABI impact (since we purposefully used unsigned int
rather than an enum type in public API, and since the new flag name
carries the same value as the reused name).

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190213232251.22647-3-eblake@redhat.com>