rpc: client: stream: notify streams of closing connection
It not done yet. As a result if we doing 'virsh console' and
libvirtd is killed we get [1] message as virsh tracks connection
status but virsh itself won't exit because it won't get notification
that stream is broken. Only after we press a key and virsh
tries to write that key code to stream we get an error. So
let's add notification that stream's underlying connection is closed.

[1] error: Disconnected from qemu:///system due to end of file

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>
Message-Id: <1550239813-870626-3-git-send-email-nshirokovskiy@virtuozzo.com>