tools: console: pass stream/fd errors to user
If console disconnected due to connection problem or problem on server
side for example it is convinient to provide the cause to the user.
If error comes from API then error is saved in virsh global variable
but as we return success from virshRunConsole if we reach waiting
stage then error is never reported. Let's track for error in event loop!

Next after failure we do a cleanup and this cleanup can overwrite
root cause. Let's save root cause and then set it to virsh error
after all cleanup is done.

Let's also add missing error reports in code.

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>
Message-Id: <1550241575-871081-3-git-send-email-nshirokovskiy@virtuozzo.com>