bhyve: support 'isa' controller for LPC
Support modeling of the 'isa' controller for bhyve. When controller is
not present in the domain XML, but domain requires it to be there (e.g.
because bootrom is used), implicitly add addition of this controller to
the command line arguments, and bind it to PCI slot 1.

PCI slot 1 is always reserved still. User can manually define any PCI
slot for the 'isa' controller, including PCI slot 1, but other devices
are not allowed to use this address.

Signed-off-by: Roman Bogorodskiy <bogorodskiy@gmail.com>
Message-Id: <20190217130405.4328-4-bogorodskiy@gmail.com>