bhyve: automatically add 'isa' controller
When domain configuration requires the 'isa' controller to be present,
automatically add it on domain post-parse stage.

Now, as this controller is always available when needed, it's not
necessary to implicitly add it to the bhyve command line, so remove
bhyveBuildLPCArgStr().

Also, make bhyveDomainDefNeedsISAController() static as it's no longer
used outside of bhyve_domain.c.

Signed-off-by: Roman Bogorodskiy <bogorodskiy@gmail.com>
Message-Id: <20190217130405.4328-5-bogorodskiy@gmail.com>