testutils: Explicitly name virTestCompare*() arguments
Currently, some arguments are called strcontent and strsrc, or
content and src or some other combination. This makes it
impossible to see at the first glance what argument is supposed
to represent 'expected' value and which one represents 'actual'
value. Rename the arguments to make it obvious.

At the same time, rework virTestCompareToULL a bit so that local
variables are named in the same fashion.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <4a449e6f7cdd96671dc81d743730c98b86addadd.1550596947.git.mprivozn@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>