conf: don't output managed attr for non-pci hostdevs
It is ignored on input for non pci hostdevs as written in docs.
I guess it would be better if the attr was disabled to be specified
in the first place instead but such behaviour is already documented.
At least let's not output it back. This will fix issue when
managed appeared on output for non pci hostdevs when it was not
specified on input.

The tests are fixed by next commands:
grep  -lRP 'hostdev.*mode=.subsystem.*type=.(?!pci).*managed' `find tests -name '*.xml'` > files
sed -i.bak -re 's/ managed=.(yes|no).//' `cat files`

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>
Message-Id: <1550828104-754472-1-git-send-email-nshirokovskiy@virtuozzo.com>