util: alloc: Remove pointless clearing of variable in AUTOPTR_FUNCs
VIR_DEFINE_AUTOPTR_FUNC defines a function which is supposed to free the
resources for the given VIR_AUTOPTR variable. Given that the cleanup
function is executed when the stack frame is being destroyed it does not
make much sense to set the pointer to NULL.

Making the inline function contain less code also decreases the
possibility that for a given type the inlining will be declined by the
compiler due to increasing code size.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <5d0750659d94f4359813b363421e1df673dda8f0.1550851418.git.pkrempa@redhat.com>