tools/virt-host-validate: Fix IOMMU check on s390x
When running virt-host-validate on an s390x host, the tool currently
warns that it is "Unknown if this platform has IOMMU support".
We can use the common check for entries in sys/kernel/iommu_groups here,
too, but it only makes sense to check it if there are also PCI devices
available. It's also common on s390x that there are no PCI devices
assigned to the LPAR, and in that case there is no need for the
PCI-related IOMMU, so without PCI devices we should simply skip this
test.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Message-Id: <1551178940-28328-1-git-send-email-thuth@redhat.com>
Reviewed-by: Bjoern Walk <bwalk@linux.ibm.com>
Reviewed-by: Boris Fiuczynski <fiuczy@linux.ibm.com>