util: XML: Introduce automatic reset of XPath's current node
Quite a few parts modify the XPath context current node to shif the
scope and allow easier queries. This also means that the node needs
to be restored afterwards.

Introduce a macro based on 'VIR_AUTOCLEAN' which adds a local structure
on the stack remembering the original node along with a function which
will make sure that the node is reset when the local structure leaves
scope.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <5c70d6a5a162c3de519340b21d0352cf967ab8a5.1551200849.git.pkrempa@redhat.com>