virDomainLoaderDefParseXML: Allow loader path to be NULL
Except not really. At least for now.

In the future, the firmware will be selected automagically.
Therefore, it makes no sense to require path to one in the domain
XML. But since it is not implemented do not really allow the path
to be NULL. Only move code around to prepare it for further
expansion.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <a7d124a6b94824f470a020b0acffd9b36d4f8bd3.1551261217.git.mprivozn@redhat.com>
Reviewed-by: Laszlo Ersek <lersek@redhat.com>