virstoragefile: identify GPFS as cluster FS
GPFS is 'IBM General Parallel File System', also called
'IBM Spectrum Scale', stalls when a process issues a lot
of fsyncs in a short timeframe.

It is better if the VM I/O is not synchronous.
It is racommended to use writeback caching policy.

In order to permit that it is added VIR_FILE_SHFS_GPFS to
the list of shared file systems in virStorageFileIsClusterFS
function.

Signed-off-by: Diego Michelotto <diego.michelotto@cnaf.infn.it>
Message-Id: <20190227140752.21801-1-diego.michelotto@cnaf.infn.it>