qemu: Factor out qemuDomainSnapshotValidate() helper
Straight code motion, coupled with changing goto into return -1
as needed. This change will be important to later patches adding
bulk redefinition (where each snapshot in a list has to meet the
same constraints), but moving it out now also makes it easier
for a more immediate bug-fix related to which states are being
validated in the next patch.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190227200428.11899-5-eblake@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>