qemu: Use virDomainSnapshotState for switch statements
Clean up the previous patch which abused a switch on virDomainState
when contrasted to a variable containing virDomainSnapshotState, by
converting the two affected switch statements to now use the right
enum.  No semantic changes now (the caller to qemuDomainSnapshotValidate
passes in a domain state rather than a snapshot state, so the new
branch is currently unreachable [will change in next patch], while
qemuDomainRevertToSnapshot can't reach the new state because of the
earlier rejection of external images).

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190227200428.11899-7-eblake@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>